% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_palette.R
\docType{data}
\name{cpr_colours}
\alias{cpr_colours}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
cpr_colours
}
\description{
Complete list of palettes
}
\keyword{datasets}
