% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPRDataViz.R
\name{theme_cpr}
\alias{theme_cpr}
\title{Title}
\usage{
theme_cpr()
}
\description{
This is a theme used with ggplot (feature of tidyverse) that the Data Viz team at the Chicago Policy Review should use in their graphics.
In order for this to work please install the font faces that are found on the CPR google drive under "Template Font Faces MUST DOWNLOAD"
}
